Addressed all PR comments.
