+++ pojo/src/main/java/com/spotify/hamcrest/pojo/MethodValueSupplierException.java
@@ -1,0 +1,40 @@
+/*-
+ * -\-\-
+ * hamcrest-pojo
+ * --
+ * Copyright (C) 2017 Spotify AB
+ * --
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * -/-/-
+ */
+
+package com.spotify.hamcrest.pojo;
+
+class MethodValueSupplierException extends RuntimeException {
+
+  private final String methodName;
+  private final String description;
+
+  public MethodValueSupplierException(String methodName, String description) {
+    this.methodName = methodName;
+    this.description = description;
+  }
+
+  public String getMethodName() {
+    return methodName;
+  }
+
+  public String getDescription() {
+    return description;
+  }
+}
