+++ pojo/src/main/java/com/spotify/hamcrest/pojo/NamedLambdaValueSupplier.java
@@ -1,0 +1,44 @@
+/*-
+ * -\-\-
+ * hamcrest-pojo
+ * --
+ * Copyright (C) 2017 Spotify AB
+ * --
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * -/-/-
+ */
+
+package com.spotify.hamcrest.pojo;
+
+import java.util.function.Function;
+
+class NamedLambdaValueSupplier<A> implements Function<A, Object> {
+
+  private final String lambdaName;
+  private final Function<A, ?> delegateFunction;
+
+  public NamedLambdaValueSupplier(String lambdaName, Function<A, ?> delegateFunction) {
+    this.lambdaName = lambdaName;
+    this.delegateFunction = delegateFunction;
+  }
+
+  @Override
+  public Object apply(A input) {
+    return delegateFunction.apply(input);
+  }
+
+  @Override
+  public String toString() {
+    return lambdaName;
+  }
+}
